Add Redux store and events slice
A Redux store and an events slice have been set up. The events slice includes actions for getting sports events and updating the app state accordingly, which are now incorporated into the Redux store. New types for EventsState and SportEvent have also been created and used in the set up.