Add selections view and top bar
This commit adds the selections feature to the application by including a new selection slice in the Redux store. In addition, a new top bar component with a navigation menu has been implemented, along with the necessary types and components related to selections. The package dependencies have been updated with the inclusion of @mui/icons-material.